Login with Cookies

+++ Web Developement 1/htdocs/resume/portfolio_single.php
@@ -36,10 +36,12 @@
 					<h2><?=$item['title']; ?></h2>
 					<p><?=nl2br($item['content']); ?></p>
 					<p>
-                        <a href="portfolio_single.php?pid=<?=$item['id']; ?>">View</a> |
-                        <a href="portfolio_edit.php?pid=<?=$item['id']; ?>">Edit</a> |
-                        <a href="portfolio_delete.php?pid=<?=$item['id']; ?>">Delete</a>
-                        <!-- we should confirm with the user before deleting -->
+                        <a href="portfolio_single.php?pid=<?=$item['id']; ?>">View</a>
+                        <?php if ($isUserLoggedIn){ ?>
+                            | <a href="portfolio_edit.php?pid=<?=$item['id']; ?>">Edit</a> |
+                            <a href="portfolio_delete.php?pid=<?=$item['id']; ?>">Delete</a>
+                            <!-- we should confirm with the user before deleting -->
+                        <?php } ?>
                     </p>
 				</article>
